Moved Folders Out
The about file and the 1.5 folders have to be in the direct subfolder of the mod.

Minor issue with major consequences
