--- FPM - Personal Additions/1.5/Defs/CodonDefs/CodonDefs.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Defs>
-
-<FPM_Additions.CodonDef Name="CodonTest">
-<defName>CodonTest</defName>
-<description>Lorem Ipsum {0}.</description>
-<iconPath>UI/Icons/Genes/Gene_KindInstinct</iconPath>
-<label> Codon Test</label>
-<transfer>false</transfer>
-<abilities>
-      <li>CodonTest</li>
-</abilities>
-<geneSet>
-    <li>KindInstinct</li>
-    <li>CodonTest</li>
-</geneSet>
-
-<removedGenes>
-    
-</removedGenes>
-</FPM_Additions.CodonDef>
-
-
-</Defs>
