--- FPM - Personal Additions/1.5/Defs/GeneDefs/Gene_Defs.xml
@@ -1,159 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Defs>
-
-<!--Genes-->
-		<!-- <GeneCategoryDef>
-		<defName>Mutation</defName>
-		<label>mutation</label>
-		<displayPriorityInXenotype>999</displayPriorityInXenotype>
-		</GeneCategoryDef>
-	
-			
-		<GeneDef>
-			<defName>FPM_ArchiteMutation</defName>
-			<label>archite mutation</label>
-			<labelShortAdj>mutating</labelShortAdj>
-			<description>Carriers of this gene will only grow stronger with age.</description>
-			<iconPath>Genes/ViewGenes</iconPath>
-			<geneClass>EBSGFramework.HediffAdder</geneClass>
-			<displayCategory>Mutation</displayCategory>
-			<minAgeActive>13</minAgeActive>
-			<modExtensions>
-				<li Class="EBSGFramework.EBSGExtension">
-					<hediffsToApplyAtAges>
-						<li>
-							<hediff>PrimaryResequence</hediff>
-							<minAge>13</minAge>
-						</li>
-					</hediffsToApplyAtAges>
-				</li>
-			</modExtensions>
-		</GeneDef>
-		
-	
-		<GeneDef>
-			<defName>FPM_Animalistic</defName>
-			<label>animalistic</label>
-			<description>Carriers of this gene have a disposition more akin to animals, boosting their effectiveness in this field.</description>
-			<iconPath>Abilities/Wildspeaker/EssenceTransfer</iconPath>
-			<displayCategory>Mutation</displayCategory>
-			<displayOrderInCategory>600</displayOrderInCategory>
-			<biostatMet>-1</biostatMet>
-			<statOffsets>
-				<HuntingStealth>0.5</HuntingStealth>
-			</statOffsets>
-			<statFactors>
-				<TameAnimalChance>1.2</TameAnimalChance>
-				<TrainAnimalChance>1.2</TrainAnimalChance>
-			</statFactors>
-				<abilities>
-				<li>FPM_Beast_Bond</li>
-				</abilities>
-		</GeneDef>
-
-	
-		<GeneDef>
-			<defName>FPM_Duelist</defName>
-			<label>duelist</label>
-			<description>Carriers of this gene are a whirlwind on the battlefield.</description>
-			<iconPath>UI/Abilities/Ravager/AAVPE_SpinCut</iconPath>
-			<displayOrderInCategory>600</displayOrderInCategory>
-			<displayCategory>Mutation</displayCategory>
-			<biostatMet>-1</biostatMet>
-			<statFactors>
-				<MeleeDamageFactor>1.25</MeleeDamageFactor>
-				<VEF_MeleeAttackSpeedFactor>1.25</VEF_MeleeAttackSpeedFactor>
-				<MeleeDodgeChance>1.25</MeleeDodgeChance>
-			</statFactors>
-
-		</GeneDef>
-
-	
-		<GeneDef>
-			<defName>FPM_HealingHands</defName>
-			<label>healing hands</label>
-			<description>Carriers of this gene generate less filth and are skilled at treating wounds.</description>
-			<iconPath>Abilities/Protector/Stabilize</iconPath>
-			<displayOrderInCategory>600</displayOrderInCategory>
-			<displayCategory>Mutation</displayCategory>
-			<biostatMet>-1</biostatMet>
-			<statFactors>
-				<FilthRate>0.5</FilthRate>
-				<MedicalTendQuality>1.25</MedicalTendQuality>
-				<MedicalSurgerySuccessChance>1.25</MedicalSurgerySuccessChance>
-			</statFactors>
-
-		</GeneDef>
-
-
-		<GeneDef>
-			<defName>FPM_Tactical</defName>
-			<label>tactical</label>
-			<description>Carriers of this gene are masters with ranged weapons.</description>
-			<iconPath>Abilities/Warlord/GuidedShot</iconPath>
-			<displayCategory>Mutation</displayCategory>
-			<displayOrderInCategory>600</displayOrderInCategory>
-			<biostatMet>-1</biostatMet>
-			<statFactors>
-				<VEF_RangeAttackSpeedFactor>1.25</VEF_RangeAttackSpeedFactor>
-				<ShootingAccuracyPawn>1.25</ShootingAccuracyPawn>
-				<AimingDelayFactor>0.75</AimingDelayFactor>
-			</statFactors>
-		</GeneDef>
-
-
-		<GeneDef>
-			<defName>FPM_Meticulous</defName>
-			<label>meticulous</label>
-			<description>Carriers of this gene become engrossed by mechanisms.</description>
-			<iconPath>Abilities/Technomancer/ReverseEngineer</iconPath>
-			<displayCategory>Mutation</displayCategory>
-			<displayOrderInCategory>600</displayOrderInCategory>
-			<biostatMet>-1</biostatMet>
-			<statFactors>
-				<WorkSpeedGlobal>1.25</WorkSpeedGlobal>
-				<ButcheryMechanoidEfficiency>1.25</ButcheryMechanoidEfficiency>
-				<RestFallRateFactor>0.75</RestFallRateFactor>
-			</statFactors>
-
-		</GeneDef>
-
-
-		<GeneDef>
-			<defName>FPM_StoneBorn</defName>
-			<label>stone born</label>
-			<description>Carriers of this gene are drawn to tunnel any way they can.</description>
-			<iconPath>Abilities/Skipmaster/Skipdoor</iconPath>
-			<displayCategory>Mutation</displayCategory>
-			<displayOrderInCategory>600</displayOrderInCategory>
-			<biostatMet>-1</biostatMet>
-			<statFactors>
-				<MiningYield>1.25</MiningYield>
-				<MiningSpeed>1.25</MiningSpeed>
-			</statFactors>
-			<ignoreDarkness>true</ignoreDarkness>
-
-		</GeneDef>
-
-
-
-		<GeneDef>
-			<defName>FPM_Gene_MultiImplanter</defName>
-			<label>MultiImplant</label>
-			<description>Lorem Ipsum.</description>
-			<iconPath>UI/Icons/Genes/Gene_XenogermReimplanter</iconPath>
-			<displayCategory>Mutation</displayCategory>
-			<displayOrderInCategory>600</displayOrderInCategory>
-			<biostatMet>-1</biostatMet>
-			<abilities>
-				<li>FPM_Ability_MultiImplant</li>
-			</abilities>
-		</GeneDef> -->
-
-
-
-
-
-
-
-</Defs>
