--- FPM - Personal Additions/1.5/Defs/ThingDefs/Codon_Defs.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Defs>
-
-<!-- Abstract base definition -->
-<CodonDef>
-    <defName>Codon_Test</defName>
-    <label>Base Heritogen</label>
-    <description>Lorem Ipsum, Heritogen bwahaha.</description>
-    <descriptionShort>Lorem Ipsum, Heritogen bwahaha.</descriptionShort>
-    <xenogene>True</xenogene> 
-
-    <!-- Caster effects-->
-    <casterRemovedGenes> 
-    </casterRemovedGenes>
-    
-    <casterAddedGenes>
-        <li>KindInstinct</li>
-    </casterAddedGenes>
-
-    <casterRemovedHediffs>
-    </casterRemovedHediffs>
-    
-    <casterAddedHediffs>
-    </casterAddedHediffs>
-
-    <!-- Target effects-->
-    <targetRemovedGenes>
-    </targetRemovedGenes>
-
-    <targetAddedGenes>
-    </targetAddedGenes>
-
-    <targetRemovedHediffs>
-    </targetRemovedHediffs>
-
-    <targetAddedHediffs>
-    </targetAddedHediffs>
-</CodonDef>
-</Defs>
