--- FPM - Personal Additions/1.5/Defs/ThingDefs/Resequence.xml
@@ -1,129 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Defs>
-	<!-- <ThingDef Name="ResequenceDrug" ParentName="MakeableDrugBase">
-	<defName>ResequenceDrug</defName>
-    <description>a cultivated collection of mechanites.</description>
-	<label>resequencing mechanites</label>
-		<graphicData>
-			<texPath>Things/Item/Drug/Penoxycyline</texPath>
-			<graphicClass>Graphic_StackCount</graphicClass>
-		</graphicData>
-		<rotatable>false</rotatable>
-		<tradeability>Sellable</tradeability>
-		<statBases>
-			<MarketValue>60</MarketValue>
-			<Mass>0.01</Mass>
-			<WorkToMake>2500</WorkToMake>
-		</statBases>
-		<techLevel>Industrial</techLevel>
-		<ingestible>
-			<foodType>Processed</foodType>
-			<drugCategory>Medical</drugCategory>
-			<outcomeDoers>
-			</outcomeDoers>
-		</ingestible>
-		<recipeMaker>
-			<recipeUsers>
-				<li>DrugLab</li>
-			</recipeUsers>
-		</recipeMaker>
-		<costList>
-			<Neutroamine>2</Neutroamine>
-			<HemogenPack>2</HemogenPack>
-		</costList>
-		<comps>
-			<li Class="CompProperties_Drug">
-				<listOrder>1000</listOrder>
-				<overdoseSeverityOffset>
-					<min>0.08</min>
-					<max>0.14</max>
-				</overdoseSeverityOffset>
-			</li>
-		</comps>
-	</ThingDef>
-
- <ThingDef Name="WarlordMechanites" ParentName="ResequenceDrug">
-	<defName>WarlordMechanites</defName>
-	<label>warlord tuned mechanites</label>
-	<description>These mechanites have been cultivated to induce development of a "Warlord"</description>
-	<ingestible>
-      <outcomeDoers>
-        <li Class="IngestionOutcomeDoer_GiveHediff">
-          <hediffDef>WarlordResequence</hediffDef>
-          <severity>1.0</severity>
-        </li>
-      </outcomeDoers>
-    </ingestible>
-</ThingDef>
-<ThingDef Name="WildspeakerMechanites" ParentName="ResequenceDrug">
-    <defName>WildspeakerMechanites</defName>
-    <label>wildspeaker tuned mechanites</label>
-    <description>These mechanites have been cultivated to induce development of a "Wildspeaker"</description>
-    <ingestible>
-        <drugCategory>Medical</drugCategory>
-        <outcomeDoers>
-            <li Class="IngestionOutcomeDoer_GiveHediff">
-                <hediffDef>WildspeakerResequence</hediffDef>
-                <severity>1.0</severity>
-            </li>
-        </outcomeDoers>
-    </ingestible>
-</ThingDef>
-<ThingDef Name="SkipmasterMechanites" ParentName="ResequenceDrug">
-    <defName>SkipmasterMechanites</defName>
-    <label>skipmaster tuned mechanites</label>
-    <description>These mechanites have been cultivated to induce development of a "Skipmaster"</description>
-    <ingestible>
-        <drugCategory>Medical</drugCategory>
-        <outcomeDoers>
-            <li Class="IngestionOutcomeDoer_GiveHediff">
-                <hediffDef>SkipmasterResequence</hediffDef>
-                <severity>1.0</severity>
-            </li>
-        </outcomeDoers>
-    </ingestible>
-</ThingDef>
-<ThingDef Name="TechnomancerMechanites" ParentName="ResequenceDrug">
-    <defName>TechnomancerMechanites</defName>
-    <label>technomancer tuned mechanites</label>
-    <description>These mechanites have been cultivated to induce development of a "Technomancer"</description>
-    <ingestible>
-        <drugCategory>Medical</drugCategory>
-        <outcomeDoers>
-            <li Class="IngestionOutcomeDoer_GiveHediff">
-                <hediffDef>TechnomancerResequence</hediffDef>
-                <severity>1.0</severity>
-            </li>
-        </outcomeDoers>
-    </ingestible>
-</ThingDef>
-<ThingDef Name="RavagerMechanites" ParentName="ResequenceDrug">
-    <defName>RavagerMechanites</defName>
-    <label>ravager tuned mechanites</label>
-    <description>These mechanites have been cultivated to induce development of a "Ravager"</description>
-    <ingestible>
-        <drugCategory>Medical</drugCategory>
-        <outcomeDoers>
-            <li Class="IngestionOutcomeDoer_GiveHediff">
-                <hediffDef>RavagerResequence</hediffDef>
-                <severity>1.0</severity>
-            </li>
-        </outcomeDoers>
-    </ingestible>
-</ThingDef>
-<ThingDef Name="ProtectorMechanites" ParentName="ResequenceDrug">
-    <defName>ProtectorMechanites</defName>
-    <label>protector tuned mechanites</label>
-    <description>These mechanites have been cultivated to induce development of a "Protector"</description>
-    <ingestible>
-        <drugCategory>Medical</drugCategory>
-        <outcomeDoers>
-            <li Class="IngestionOutcomeDoer_GiveHediff">
-                <hediffDef>ProtectorResequence</hediffDef>
-                <severity>1.0</severity>
-            </li>
-        </outcomeDoers>
-    </ingestible>
-</ThingDef> -->
-
-</Defs>
