--- FPM - Personal Additions/1.5/Defs/XenotypeDefs/Sephari.xml
@@ -1,41 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Defs>
-  <XenotypeDef>
-    <defName>GegeXT_Sephari</defName>
-    <label>Sephari</label>
-    <description>A race of fallen warriors, each lineage contains limitless possibilites.</description>
-    <descriptionShort></descriptionShort>
-    <iconPath>UI/Icons/Xenotypes/Custom/Sanguophage_4</iconPath>
-    <inheritable>true</inheritable>
-    <genes>
-		<li MayRequire="gege.exportedxenotypestoragemodfolder">FPM_ArchiteMutation</li>
-		<li MayRequire="vanillaracesexpanded.lycanthrope">VRE_MoveSpeed_ExtremelyQuick</li>
-		<li>DiseaseFree</li>
-		<li MayRequire="vanillaracesexpanded.archon">VRE_InnatePsylink</li>
-		<li MayRequire="sarg.alphagenes">AG_Invisibility</li>
-		<li MayRequire="sarg.alphagenes">AG_DangerousRaidPresence</li>
-		<li MayRequire="vanillaracesexpanded.genie">VRE_BlueBlood</li>
-		<li>RobustDigestion</li>
-		<li>WoundHealing_Slow</li>
-		<li>Superclotting</li>
-		<li MayRequire="sarg.alphagenes">AG_RegenerationMinor</li>
-		<li MayRequire="vanillaracesexpanded.archon">VRE_PsychicAbility_MoreExtreme</li>
-		<li MayRequire="vanillaracesexpanded.hussar">VREH_Unconstrained</li>
-		<li>Robust</li>
-		<li MayRequire="sarg.alphagenes">AG_ReducedFertile</li>
-		<li MayRequire="cyanobot.cyanobotsgenes">Hypercarnivore</li>
-		<li MayRequire="cyanobot.cyanobotsgenes">CYB_ExpansiveStomach</li>
-		<li MayRequire="cyanobot.cyanobotsgenes">CYB_Precocious</li>
-		<li MayRequire="cyanobot.cyanobotsgenes">CYB_FastMaturation</li>
-		<li MayRequire="petre27.biotechbalancedcapacitorgenes">Balanced_Capacity_Boost_Tier_4</li>
-		<li MayRequire="vanillaracesexpanded.archon">VRE_DreadHair</li>
-		<li>Body_Standard</li>
-		<li MayRequire="cyanobot.cyanobotsgenes">CYB_Hair_Pearl</li>
-		<li MayRequire="sarg.alphagenes">AG_Skin_DeepPurple</li>
-		<li MayRequire="vanillaracesexpanded.lycanthrope">VRE_PackMentality</li>
-		<li>Learning_Fast</li>
-		<li MayRequire="vanillaracesexpanded.fungoid">VRE_SkillDegradation</li>
-		<li MayRequire="cyanobot.cyanobotsgenes">CYB_Offspring_GegeXT_Sephari</li>
-	</genes>
-  </XenotypeDef>
-</Defs>
