--- FPM - Personal Additions/1.5/Source/FPM - Personal Additions/FPM - Personal Additions.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.10.35122.118
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FPM - Personal Additions", "FPM - Personal Additions.csproj", "{12FE8982-48CD-4881-B4A6-BCB9A6167DFB}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{12FE8982-48CD-4881-B4A6-BCB9A6167DFB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{12FE8982-48CD-4881-B4A6-BCB9A6167DFB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{12FE8982-48CD-4881-B4A6-BCB9A6167DFB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{12FE8982-48CD-4881-B4A6-BCB9A6167DFB}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {BFC3A142-9DA2-4E61-8822-39751240929F}
-	EndGlobalSection
-EndGlobal
