--- FPM - Personal Additions/1.5/Source/FPM - Personal Additions/FPM_Utilites.cs
@@ -1,19 +1,0 @@
-﻿using RimWorld;
-using System.Collections.Generic;
-using System.Text;
-using Verse;
-
-namespace FPM_Additions
-{
-
-
-   
-    public static class FPM_Utilites
-    {
-
-
-        
-    }
-
-
-}
