--- FPM - Personal Additions/About/About.xml
@@ -1,19 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ModMetaData>
-  <name>FPM Personal Additions</name>
-  <author>False</author>
-  <packageId>false.personaladditions</packageId>
-  <supportedVersions>
-    <li>1.5</li>
-  </supportedVersions>
-  <description>Custom Changes.</description>
-  <modDependencies>
-    <li>
-      <packageId>Ludeon.RimWorld.Biotech</packageId>
-      <displayName>Biotech</displayName>
-    </li>
-  </modDependencies>
-  <loadAfter>
-    <li>Ludeon.RimWorld.Biotech</li>
-  </loadAfter>
-</ModMetaData>
